interpoint/pyproject.toml
[project]
name = "interactvlm-discrete"
version = "0.1.0"
description = "Discrete Human-Object Interaction Contact Prediction from Single RGB Image"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = ["human-object-interaction", "contact-prediction", "vision-language-model", "deep-learning"]

dependencies = [
    # Core (LOCKED to a compatible pair)
    "torch==2.4.*",
    "torchvision==0.19.*",
    "numpy>=1.24.0",
    "scipy>=1.10.0",

    # Vision Transformer & VLM
    "transformers>=4.36.0",
    "peft>=0.7.0",
    "accelerate>=0.25.0",

    # 3D Rendering
    "trimesh>=4.0.0",
    "smplx>=0.1.28",

    # Image processing
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "albumentations>=1.3.0",

    # Configuration
    "omegaconf>=2.3.0",

    # Logging & Monitoring
    "tensorboard>=2.14.0",
    "tqdm>=4.65.0",

    # Utilities
    "einops>=0.7.0",
    "timm>=0.9.0",

    # Metrics
    "scikit-learn>=1.3.0",

    # YAML
    "pyyaml>=6.0",
    "bitsandbytes>=0.46.1"
]

[project.optional-dependencies]
wandb = [
    "wandb>=0.16.0",
]
smplx = [
    "smplx>=0.1.28",
]
full = [
    "wandb>=0.16.0",
    "smplx>=0.1.28",
    "pyrender>=0.1.45",
    "hydra-core>=1.3.0",
]

[project.scripts]
ivd-train = "train:main"
ivd-inference = "inference:main"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "data", "utils"]

[tool.uv]
# Default index for most packages (fast mirror)
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple/"

# PyTorch official CUDA wheels (better compatibility for ops like torchvision::nms)
[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]
